requirements.txt
# Freqtrade (install separately: pip install freqtrade)
# TA-Lib (requires system library: https://ta-lib.github.io/ta-lib-python/)

# Data contracts
pydantic>=2.0,<3.0

# News ingestion
feedparser>=6.0,<7.0

# Sentiment analysis
transformers>=4.30,<5.0
torch>=2.0

# LLM decision gate
openai>=1.0,<2.0

# Environment variables
python-dotenv>=1.0,<2.0

# Testing
pytest>=7.0,<9.0
pytest-mock>=3.0,<4.0
